Filter out static fields while inspecting
