+++ src/main/java/pl/wavesoftware/utils/stringify/impl/InspectingField.java
@@ -4,7 +4,7 @@ package pl.wavesoftware.utils.stringify.impl;
  * @author <a href="mailto:krzysztof.suszynski@coi.gov.pl">Krzysztof Suszynski</a>
  * @since 27.04.18
  */
-public interface InspectingField {
+interface InspectingField {
   boolean shouldInspect();
   boolean showNull();
 }
