+++ src/main/java/pl/wavesoftware/utils/stringify/impl/InspectingFieldFactory.java
@@ -2,7 +2,6 @@ package pl.wavesoftware.utils.stringify.impl;
 
 import lombok.RequiredArgsConstructor;
 import pl.wavesoftware.utils.stringify.configuration.BeanFactory;
-import pl.wavesoftware.utils.stringify.configuration.DisplayNull;
 import pl.wavesoftware.utils.stringify.configuration.InspectionPoint;
 import pl.wavesoftware.utils.stringify.configuration.Mode;
 
@@ -27,30 +26,4 @@ final class InspectingFieldFactory {
     }
   }
 
-  @RequiredArgsConstructor
-  private class InspectingFieldImpl implements InspectingField {
-    private final InspectionPoint inspectionPoint;
-    private final InspectFieldPredicate predicate;
-
-    @Override
-    public boolean shouldInspect() {
-      return technically() && predicate.shouldInspect(inspectionPoint);
-    }
-
-    private boolean technically() {
-      return !inspectionPoint.getField().isEnumConstant()
-        && !inspectionPoint.getField().isSynthetic();
-    }
-
-    @Override
-    public boolean showNull() {
-      DisplayNull displayNull = inspectionPoint.getField()
-        .getAnnotation(DisplayNull.class);
-      if (displayNull != null) {
-        return displayNull.value();
-      } else {
-        return DisplayNull.BY_DEFAULT;
-      }
-    }
-  }
 }
