+++ src/test/java/pl/wavesoftware/utils/stringify/Planet.java
@@ -5,12 +5,17 @@ import pl.wavesoftware.utils.stringify.configuration.DisplayNull;
 import pl.wavesoftware.utils.stringify.configuration.DoNotInspect;
 import pl.wavesoftware.utils.stringify.configuration.Inspect;
 
+import java.io.Serializable;
+
 /**
  * @author <a href="krzysztof.suszynski@wavesoftware.pl">Krzysztof Suszyński</a>
  * @since 2018-04-18
  */
 @Data
-abstract class Planet {
+abstract class Planet implements Serializable {
+
+  private static final long serialVersionUID = 20180430201529L;
+
   @Inspect
   private String name;
   @Inspect
